chore: Move FormattedDate and TimeAgo into common
